Revert "Fix a compilation error on x86 systems by converting TotalSwap and FreeSwap to uint64."
This reverts commit d368400dc631611d66d9fb5279a9332a80063e83.

Reverting due to missing CLA

+++ sigar_linux.go
@@ -88,8 +88,8 @@ func (self *Swap) Get() error {
 		return err
 	}
 
-	self.Total = uint64(sysinfo.Totalswap)
-	self.Free = uint64(sysinfo.Freeswap)
+	self.Total = sysinfo.Totalswap
+	self.Free = sysinfo.Freeswap
 	self.Used = self.Total - self.Free
 
 	return nil
